Extend CoapOption by two Options: CoapOptionLength & CoapOptionDelta
The small delta and length are partly hints to check for their respective extended versions. In case of an extension one can either save the extended version and reconstruct the small version if needed or save the simple and extended version if required. The second case avoids reconstruction and resembles the model more closely.

getDelta and getLength as part of the CoapExtractionService should be split up into smaller functions at this point.